refactor: move dc ui components and rename usage
