+++ src/components/dc-ui/DcRecruitForm.vue
@@ -532,7 +532,7 @@
                   {{ f.label }}
                 </div>
                 <div class="dc-pad-16">
-                  <Uploader
+                  <dc-uploader
                     v-model="uploaderModels[f.name]"
                     :multiple="(f.upload?.maxCount || 1) > 1"
                     :max-count="f.upload?.maxCount || 1"
@@ -554,7 +554,7 @@
                   :title-style="cellTitleStyle(f)"
                 />
                 <div class="dc-pad-16">
-                  <Uploader
+                  <dc-uploader
                     v-model="uploaderModels[f.name]"
                     :multiple="(f.upload?.maxCount || 1) > 1"
                     :max-count="f.upload?.maxCount || 1"
