+++ src/components/index.js
@@ -1,14 +1,14 @@
-import Dict from './Dict.vue';
 import LanguageSelector from './LanguageSelector.vue';
-import RecruitForm from './RecruitForm.vue';
 import ScrollAwareNavBar from './ScrollAwareNavBar.vue';
-import Uploader from './Uploader.vue';
+import DcDict from './dc-ui/DcDict.vue';
+import DcRecruitForm from './dc-ui/DcRecruitForm.vue';
+import DcUploader from './dc-ui/DcUploader.vue';
 import DcPagination from './dc-ui/DcPagination.vue';
 
 export function registerComponents(app) {
-  app.component('Uploader', Uploader);
-  app.component('RecruitForm', RecruitForm);
-  app.component('Dict', Dict);
+  app.component('DcUploader', DcUploader);
+  app.component('DcRecruitForm', DcRecruitForm);
+  app.component('DcDict', DcDict);
   app.component('LanguageSelector', LanguageSelector);
   app.component('ScrollAwareNavBar', ScrollAwareNavBar);
   app.component('DcPagination', DcPagination);
