+++ src/views/recruit/campus/apply.vue
@@ -1,7 +1,7 @@
 <!-- src/views/recruit/campus/apply.vue -->
 <template>
   <img class="banner" :src="$assetUrl('/images/recruit/campus/apply/banner.svg')" alt="banner" />
-  <RecruitForm
+  <dc-recruit-form
     v-model="form"
     :schema="schema"
     submit-text="提交"
